==> 0 all_data_merged.ipynb <==
import warnings
warnings.simplefilter(action='ignore')

import pandas as pd
import numpy as np
import datetime

datapath='Processed_Data/city_compare.h5'
db=pd.HDFStore(datapath)
db.keys()

airports=db['airports']
amap=db['amap']
hdl=db['hdl']
heytea=db['heytea']
hotel=db['hotel']
subway=db['subway']
movie=db['movie']
bdcom=db['BDcom']

db.close()

city_popu=pd.read_hdf('./Processed_Data/citypopu.h5',key='raw2')
city_popu['城市']=city_popu['城市'].str.replace('市',"")
city_popu['常住人口']=pd.to_numeric(city_popu['常住人口'])

# citytier=pd.read_clipboard()
citytier.to_hdf('./Processed_Data/citypopu.h5',key='citytier')

df_covid=pd.read_hdf('Processed_Data/Baidu_covid_progress.h5',key='raw')
df_covid=df_covid[['city','peak','progress_2']]
df_covid.columns=['城市','峰值日','covid_progess']
df_covid['covid_progess']=pd.to_numeric(df_covid['covid_progess'])/100
df_covid.head(2)

all=city_popu.merge(df_covid,how='outer',on='城市').merge(amap.drop(columns='date'),how='outer',on='城市').merge(airports.drop(columns='date'),how='outer',on='城市').merge(subway.drop(columns='date'),how='outer',on='城市').merge(bdcom.drop(columns='date'),how='outer',on='城市').merge(heytea.drop(columns='date'),how='outer',on='城市').merge(movie,how='outer',on='城市').merge(hdl.drop(columns='date'),how='outer',on='城市').merge(hotel.drop(columns='date'),how='outer',on='城市')


all['N']=all.count(axis=1)
all=all[all['covid_progess']>=0]

all=all.sort_values(by='N',ascending=False).head(1000).sort_values(by='常住人口',ascending=False).head(1000).drop(columns='N')
all.drop(columns='峰值日').to_clipboard(index=False)


d1=pd.read_hdf('Processed_Data/tablehead.h5','all')
# d1.to_hdf('Processed_Data/tablehead.h5','all')
d1.merge(all,on='城市',how='left').to_clipboard(index=False)

%load_ext autoreload
%autoreload 2
from colored_list import dafeng_city,jidui_city,riqi
all[all['城市'].isin(['上海'])]
