==> 2.2 AirportsData.ipynb <==
# 1.Daily covid cases

import warnings
warnings.simplefilter(action='ignore')

import pandas as pd
import numpy as np
import datetime
# df=pd.read_excel(r'DataSource/airports.xlsx')
# df=df.drop(columns=['Unit'])
df=pd.read_excel(r'DataSource/flights.xlsx')
df=df.drop(columns=['更新时间','来源','频率','单位'])
df=df.dropna(subset=['指标名称'])
c1=df['指标名称'].str.contains('数据来源')
df=df[~c1]
df['指标名称']=df['指标名称'].str.replace('机场进出港航班:','')
df[['城市', '机场']] = df['指标名称'].str.split("/", expand = True)
df=df.drop(columns=['指标名称'])
cols=list(df.columns[-2:])+sorted(list(df.columns[:-2]),reverse=False)
df=df[cols]

df.to_hdf('./Processed_Data/airports.h5',key='raw')
df

ap=pd.read_hdf('./Processed_Data/airports.h5',key='raw')
ap=ap.groupby(by='城市').sum().reset_index()
lastcol=ap.columns[-1]
ap=ap.sort_values(by=lastcol,ascending=False).head(50)
ap=pd.melt(ap,id_vars=['城市'],var_name='date',value_name='air_traffic')
ap=ap[ap['air_traffic']>0]

ap['date'] = pd.to_datetime(ap['date']).dt.date
ap=ap.sort_values(by=['城市','date'],ascending=True)
ap['date']=ap['date'].astype('str')

def getdt(d):
    return str(d)[-5:]
ap['dt']=ap['date'].apply(getdt)

ap['ap_last7d_mean']=ap.groupby('城市')['air_traffic'].transform(lambda x: x.rolling(7, 1).mean())
ap['ap_WoW']=ap['ap_last7d_mean']/ap.groupby('城市')['ap_last7d_mean'].transform(lambda x: x.shift(periods=7)) -1

c1=ap['date'].str.contains('2019')
ap_19=ap[c1]
ap_19=ap_19[['城市', 'dt','ap_last7d_mean']]
ap_19=ap_19.rename(columns={'ap_last7d_mean':'ap_19_rolling_mean'})

c2=ap['date'].str.contains('2022|2023')
ap_22=ap[c2]
ap_22=ap_22.merge(ap_19,how='left',on=['城市','dt'])

ap_22['ap_status']=ap_22['ap_last7d_mean']/ap_22['ap_19_rolling_mean']

ap_22.tail(3)

# TY 机场 vs. 2019
%load_ext autoreload
%autoreload 2
from colored_list import dafeng_city,jidui_city,riqi
riqi='2023-01-08'

df_covid=pd.read_hdf('Processed_Data/Baidu_covid_progress.h5',key='raw')
o=ap_22.merge(df_covid,on=['城市'],how='left')

o=o[o['ap_status']<1.5][o['date']==riqi][['城市','progress_2','ap_status']] 

o['size']=np.nan
o['group']='其他'

c1=o['城市'].isin(dafeng_city)
o.loc[c1,'group']='本波疫情基本结束'

o=o.sort_values(by='group',ascending=False)
o.to_clipboard(index=False,header=None)

d=ap_22[ap_22['date']==riqi][['城市','date','ap_status']].reset_index(drop=True).rename(columns={'ap_status':'airports'})
d['date']=pd.to_datetime(d['date'])
d.to_hdf('Processed_Data/city_compare.h5',key='airports')

print(riqi)
print(o)

# TY 机场数据 WoW
# %load_ext autoreload
# %autoreload 2
# from colored_list import dafeng_city,jidui_city,riqi

df_covid=pd.read_hdf('Processed_Data/Baidu_covid_progress.h5',key='raw')
o=ap_22.merge(df_covid,on=['城市'],how='left')

o=o[o['date']==riqi][['城市','progress_2','ap_WoW']] 

o['size']=np.nan
o['group']='其他'

c1=o['城市'].isin(dafeng_city)
o.loc[c1,'group']='本波疫情基本结束'

o=o.sort_values(by='group',ascending=False)

o.to_clipboard(index=False,header=None)
print(riqi)

# TY 机场数据 WoW vs. GDP
# %load_ext autoreload
# %autoreload 2
# from colored_list import dafeng_city,jidui_city,riqi

city_popu=pd.read_hdf('Processed_Data/citypopu.h5',key='raw2')
o=ap_22.merge(city_popu,on=['城市'],how='left')

o=o[o['date']==riqi][['城市','GDP','ap_WoW']] 

o['size']=np.nan
o['group']='其他'

c1=o['城市'].isin(dafeng_city)
o.loc[c1,'group']='本波疫情基本结束'

o=o.sort_values(by='group',ascending=False)

o.to_clipboard(index=False,header=None)
print(riqi)

# TY 机场数据 as of 19 vs. GDP
# %load_ext autoreload
# %autoreload 2
# from colored_list import dafeng_city,jidui_city,riqi

city_popu=pd.read_hdf('Processed_Data/citypopu.h5',key='raw2')
o=ap_22.merge(city_popu,on=['城市'],how='left')

o=o[o['date']==riqi][o['ap_status']<1.5][['城市','GDP','ap_status']] 

o['size']=np.nan
o['group']='其他'

c1=o['城市'].isin(dafeng_city)
o.loc[c1,'group']='本波疫情基本结束'

o=o.sort_values(by='group',ascending=False)

o.to_clipboard(index=False,header=None)
print(riqi)

# d1=pd.read_clipboard()
# d1.to_hdf('Processed_Data/tablehead.h5','airports')
d1=pd.read_hdf('Processed_Data/tablehead.h5','airports')

apo=pd.read_hdf('./Processed_Data/airports.h5',key='raw')
cols=[i for i in apo.columns if i >'2023-01-04']

d1.merge(apo[cols],on=['城市','机场'],how='left').drop(columns=['城市','机场']).to_clipboard(index=False)
