==> 3.0 BaiduComsuption.ipynb <==
import requests,time,json

j=0
d=json.load(open('OneTimeSource\data.json', 'r'))
d=[i for key in d.keys() for i in d[key]]
for i in d[j:]:
    params = {
    'sv': '',
    'cuid': '',
    'os': '',
    'zid': '',
    'business': 'return_work',
    'city_id': i['id'],
    'tab_key': 'consume',
    'module_name': 'tendency',
    'days': '90',
    'callback': 'jsonp1673408387540',
    }
    response = requests.get('https://ugc.map.baidu.com/diymap/index_map/main', params=params)
    r=response
    print(j,i['id'],i['name'],r)
    j+=1
    with open('OneTimeSource\BaiduRecovery\%s.json' % i['name'], 'w') as f:
        json.dump(r.text, f)
    

import warnings
warnings.simplefilter(action='ignore')
import pandas as pd
import json,os
dir='./OneTimeSource/BaiduRecovery/'
out=pd.DataFrame()
for f in os.listdir(dir):
    text=open(dir+f,'r',encoding='utf-8').read()[1:-1]
    d=text.replace("jsonp1673408387540(","").replace('''\\"''','''"''')[:-1]
    d=json.loads(d)
    name=d['data']['modules']['tendency']['selected_city']['name']
    points=d['data']['modules']['tendency']['points']
    df= pd.DataFrame([[i['show_name'],i['value']] for i in points],columns=['date','baidu_consumption'])
    df['城市']=name.replace('市','')
    df['date']=pd.to_datetime(df['date'])
    out=pd.concat([out,df])
out.to_hdf(r'Processed_Data/baidu_consumption.h5',key='0112')
out.tail(2)

#set baseline
d=out.copy()
c1=d.date.astype('str')<='2022-11-28'
d_base=d[c1].groupby(by='城市').mean().reset_index().rename(columns={'baidu_consumption':'index_base'})
d=d.merge(d_base,on=['城市'],how='left')

#set 无疫情baseline
citydate=pd.read_hdf('Processed_Data/amap.h5',key='noCOVIDdf')
c1=d.date.astype('str')<='2022-11-28'
d['城市date']=d['城市']+d['date'].astype('str')
c2=d['城市date'].isin(citydate['城市date'])
d_basenoCOVID=d[c1&c2]
d_basenoCOVID['WeekDay']=pd.to_datetime(d_basenoCOVID['date']).dt.strftime('%A')
def workday(s):
    if s in ['Saturday','Sunday']:
        return 'Weekend'
    else:
        return 'Workday'
d_basenoCOVID['Workday']=d_basenoCOVID['WeekDay'].apply(workday)
d_basenoCOVID=pd.pivot_table(d_basenoCOVID,index='城市',columns='Workday',values='baidu_consumption',aggfunc='mean').reset_index()
d_basenoCOVID=d_basenoCOVID[(d_basenoCOVID['Weekend']>0) & (d_basenoCOVID['Workday']>0)]
d_basenoCOVID['index_base_nocovid']=(d_basenoCOVID['Weekend']*2+d_basenoCOVID['Workday']*5)/7
d_basenoCOVID=d_basenoCOVID[['城市','index_base_nocovid']]

d=d.merge(d_basenoCOVID,on=['城市'],how='left')

#cal wow
d['index_last7d_mean']=d.groupby('城市')['baidu_consumption'].transform(lambda x: x.rolling(7, 7).mean())
d['index_change_WoW']=(d['index_last7d_mean']/d.groupby('城市')['index_last7d_mean'].transform(lambda x: x.shift(periods=7)))-1

#cal as of 
d['asof2211']=d['index_last7d_mean']/d['index_base_nocovid'] # index_base
d.tail(2)

#总体恢复情况 as of 无疫情 
c1=d['date']>='2022-12-01'
c2=d['index_base_nocovid']>0

t=d[c1 & c2].groupby('date').sum().reset_index()
t['bd']=t['index_last7d_mean']/t['index_base_nocovid']
t[['date','bd']].to_clipboard(index=False)

# 人均收入 vs.bd index WoW
import numpy as np
riqi='2023-01-16'

bd=d.copy()
city_popu=pd.read_hdf('Processed_Data/citypopu.h5',key='raw2')
bd=bd.merge(city_popu,on='城市',how='left')
bd['date']=bd['date'].astype('string') 

bd=bd[bd['date']==riqi].sort_values(by='常住人口',ascending=False)
bd=bd[bd['index_change_WoW']<3].head(70).reset_index()

o=bd[['城市',  '人均可支配收入','index_change_WoW']]

o['size']=np.nan
o['group']='其他'

o=o.sort_values(by='group',ascending=False)

o.to_clipboard(index=False,header=None)
print(riqi,'wow')
o.shape

# 人均收入 vs. bd index as of 2022

bd=d.copy()
city_popu=pd.read_hdf('Processed_Data/citypopu.h5',key='raw2')
bd=bd.merge(city_popu,on='城市',how='left')
bd['date']=bd['date'].astype('string') 

bd=bd[bd['date']==riqi].sort_values(by='常住人口',ascending=False)
bd=bd[bd['asof2211']<3].head(100).reset_index()

o=bd[['城市',  '人均可支配收入','asof2211']]

o['size']=np.nan
o['group']='其他'

o=o.sort_values(by='group',ascending=False)

td=d[d['date']==riqi][['城市','date','asof2211']].reset_index(drop=True).rename(columns={'date':'date','asof2211':'BDcomsu'})
td['date']=pd.to_datetime(td['date'])
td.to_hdf('Processed_Data/city_compare.h5',key='BDcom')

o.to_clipboard(index=False,header=None)
print(riqi,'asof')
o.shape

# d1=pd.read_clipboard()
# d1.to_hdf('Processed_Data/tablehead.h5','BDxiaofei')

# d2=pd.pivot_table(d,index=['城市'],columns=['date'],values='baidu_consumption',aggfunc='mean')
# d2.sort_values(by=d2.columns[-1],ascending=False).to_clipboard()

d1=pd.read_hdf('Processed_Data/tablehead.h5','BDxiaofei')
d2=pd.pivot_table(d[d['date'].astype('str')>='2023-01-12'],index='城市',columns='date',values='baidu_consumption',aggfunc='mean').reset_index()
d1.merge(d2,left_on='城市',right_on='城市',how='left').drop(columns=['城市']).to_clipboard(index=False)
