==> 3.4.movie.ipynb <==
import warnings
warnings.simplefilter(action='ignore')

import pandas as pd
import numpy as np
import datetime

# df=pd.read_clipboard()
# df.to_hdf(r'Processed_Data/movie.h5','1230')

df=pd.read_hdf(r'Processed_Data/movie.h5','1230')

df['MovieRecovery']=pd.to_numeric(df['MovieRecovery'].str.replace('%',''))/100
df['MovieWoW']=pd.to_numeric(df['MovieWoW'].astype('str').str.replace('%',''))/100
df['MoviePpts']=pd.to_numeric(df['MoviePpts'].astype('str').str.replace('%',''))/100


d=df[['城市','MovieRecovery']].reset_index(drop=True).rename(columns={'MovieRecovery':'movie'})
d.to_hdf('Processed_Data/city_compare.h5',key='movie')

df.head(2)

# # covid rank vs. Movie vs. 2022

# %load_ext autoreload
# %autoreload 2
# from colored_list import dafeng_city,jidui_city,riqi


# df_covid=pd.read_hdf('Processed_Data/Baidu_covid_progress.h5',key='raw')
# o=df.merge(df_covid,on='城市',how='left')

# o=o[['城市', 'progress_2','MovieRecovery']]


# o['size']=np.nan
# o['group']='其他'

# c1=o['城市'].isin(dafeng_city)
# o.loc[c1,'group']='本波疫情基本结束'

# o=o.sort_values(by='group',ascending=False) 
# o.to_clipboard(index=False,header=None)

# print('vs. 2022')

# 人均收入 vs. Movie WoW

%load_ext autoreload
%autoreload 2
from colored_list import dafeng_city,jidui_city,riqi


city_popu=pd.read_hdf('Processed_Data/citypopu.h5',key='raw2')

o=df.merge(city_popu,on='城市',how='left')

o=o[['城市', '人均可支配收入','MovieWoW']]

o['size']=np.nan
o['group']='其他'

c1=o['城市'].isin(dafeng_city)
o.loc[c1,'group']='本波疫情基本结束'

o=o.sort_values(by=['group','人均可支配收入'],ascending=False) 
o.to_clipboard(index=False,header=None)

print('WoW')

# 人均收入 vs. Movie vs. 2022

%load_ext autoreload
%autoreload 2
from colored_list import dafeng_city,jidui_city,riqi


city_popu=pd.read_hdf('Processed_Data/citypopu.h5',key='raw2')

o=df.merge(city_popu,on='城市',how='left')

o=o[['城市', '人均可支配收入','MovieRecovery']]

o['size']=np.nan
o['group']='其他'

c1=o['城市'].isin(dafeng_city)
o.loc[c1,'group']='本波疫情基本结束'

o=o.sort_values(by=['group','人均可支配收入'],ascending=False) 
o.to_clipboard(index=False,header=None)

print('vs. 2022')
