==> 4.3.binyiguan_index.ipynb <==
import warnings
warnings.simplefilter(action='ignore')

import pandas as pd
import numpy as np
import datetime
city_popu=pd.read_hdf('./Processed_Data/citypopu.h5',key='raw2')
city_popu['城市']=city_popu['城市'].str.replace('市',"")
city_popu['常住人口']=pd.to_numeric(city_popu['常住人口'])

import os
dfo=pd.DataFrame()

for f in os.listdir('./DataSource/baiduindex_binyiguan/'):
    print(f)
    d=pd.read_csv('./DataSource/baiduindex_binyiguan\%s' % f)
    dfo=pd.concat([dfo,d],axis=0)
dfo=dfo.merge(city_popu,left_on='cityName',right_on='城市',how='left')
dfo['startDate']=pd.to_datetime(dfo['startDate'])
dfo

dfo[dfo['startDate'].astype('str')>'2023-01-02']

c1=dfo['startDate']>'2022-11-01'
by=dfo[c1][['cityName','startDate','all','常住人口']]

# datelist=by['startDate'].drop_duplicates().sort_values()[-20:]

def get_status(x):
    now=float(x['now'])
    max=float(x['max'])
    per=now/max
    if x['startDate']<=x['max_date']:
        return '0 未达峰'
    else:
        if per >0.8:
            return '1 下降至 80-100%'
        if 0.5<per <=0.8:
            return '2 下降至 50-80%'
        if 0.3 <per <=0.5:
            return '3 下降至 20-50%'
        if per<=0.3:
            return '4 下降至 20%以下'


# c2=by['startDate']=='2022-12-28'
d1=by.rename(columns={'all':'now'})
d2=by.groupby(by='cityName')['all'].max().reset_index().rename(columns={'all':'max'})
d2=d2.merge(d1.sort_values(by=['cityName','startDate']).drop_duplicates(subset=['cityName','now']),left_on=['cityName','max'],right_on=['cityName','now'],how='left').rename(columns={'startDate':'max_date'})
df=d1.merge(d2[['cityName', 'max', 'max_date']],on='cityName',how='left')

df['status']=df.apply(lambda x:get_status(x),axis=1)

pv_value=pd.pivot_table(df.drop_duplicates(), values='常住人口', index=['startDate'],columns=["status"], aggfunc=np.sum)
pv_value=pv_value.div(pv_value.sum(axis=1), axis=0)

pv_value['0 sum']=1
pv_value=pv_value.fillna(0)
pv_value=pv_value.sort_index(axis=1)

pv_value.tail(15).to_clipboard(header=None)

# d1=pd.read_clipboard()
# d1.to_hdf('Processed_Data/tablehead.h5','BYG')

d1=pd.read_hdf('Processed_Data/tablehead.h5','BYG')
d2=pd.pivot_table(dfo[dfo['startDate'].astype('str')>='2023-01-11'],index='cityName',columns='startDate',values='all',aggfunc='mean')
d1.merge(d2,how='left',on='cityName').drop(columns='cityName').to_clipboard(index=False)
